114- feat(QrCodeScanner): Create a file to handle with all possible feedbacks on Qr code scanner page
- Move qr_code_scanner_page.dart to a folder qr_code_scanner
- Add a onPop voidcallback on UDialogSingleButton
- Handle better with dialogs and time to do a new scan of a qr code
